pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "wdb.version.__version__"}
readme = {file = ["README.md"], content-type = "text/markdown"}

[project]
name = "windows-desktop-builds"
description = "Retrieves Windows Desktop Builds"
requires-python = ">=3.12"
authors = [
  {email = "12127434+whoDoneItAgain@users.noreply.github.com"},
  {name = "whoDoneItAgain"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dynamic = [
    "version",
    "readme"
]

dependencies = [
    "beautifulsoup4==4.12.3",
    "packaging==24.2",
    "pyyaml==6.0.2",
    "requests==2.32.3",
]

[project.scripts]
wdb = "wdb.__main__:main"

[tool.setuptools.package-data]
amc = ["data/**/*.yaml"]

[project.urls]
Homepage = "https://github.com/whodoneitagain/windows-desktop-builds"
